add group , contributions struct and values
