+++ backend/services/contributions.go
@@ -1,0 +1,28 @@
+package services
+
+import "time"
+
+// get group total amount
+// get user total savings
+// get user total loans
+// get user dividends
+// get user contribution records to diplay
+// generate report and download
+// initiate payment >> update db
+
+type Contribution struct {
+	ID       int       `json:"id"`
+	MemberID int       `json:"member_id"`
+	Amount   float64   `json:"amount"`
+	Date     time.Time `json:"date"`
+}
+
+func GetContributions() ([]Contribution, error) {
+	// TODO: Implement database query to fetch contributions
+	return []Contribution{}, nil
+}
+
+func AddContribution(contribution Contribution) error {
+	// TODO: Implement database insert for new contribution
+	return nil
+}
